backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
bcrypt
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==23.0
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
utils==1.0.1
bcrypt
Pillow